# packages/models/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/friends": {
      "get": {
        "operationId": "friends",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List all friends",
        "tags": [
          "friends"
        ]
      },
      "post": {
        "operationId": "createFriend",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFriendDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendDto"
                }
              }
            }
          }
        },
        "summary": "Create a friend",
        "tags": [
          "friends"
        ]
      }
    },
    "/friends/upcoming/sections": {
      "get": {
        "operationId": "upcomingFriendSections",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Match against friend name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "required": false,
            "in": "query",
            "schema": {
              "default": "month",
              "type": "string",
              "enum": [
                "month",
                "age",
                "year",
                "none"
              ]
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 30,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUpcomingSectionsDto"
                }
              }
            }
          }
        },
        "summary": "Paginated, grouped upcoming birthdays for the timeline's infinite scroll",
        "tags": [
          "friends"
        ]
      }
    },
    "/friends/birthdays-by-month": {
      "get": {
        "operationId": "birthdaysByMonth",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "description": "Calendar month, 1-12",
            "schema": {
              "example": 8,
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Match against friend name",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BirthdaysByMonthDto"
                }
              }
            }
          }
        },
        "summary": "Friends whose birthday falls in a given calendar month",
        "tags": [
          "friends"
        ]
      }
    },
    "/friends/search": {
      "get": {
        "operationId": "searchFriends",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Match against friend name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedFriendsDto"
                }
              }
            }
          }
        },
        "summary": "Paginated, name-filtered friend search (used by the onboarding connections picker)",
        "tags": [
          "friends"
        ]
      }
    },
...
    "/friends/{id}": {
      "get": {
        "operationId": "friend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendDto"
                }
              }
            }
          }
        },
        "summary": "Get a friend by id or by the @handle of its associated account",
        "tags": [
          "friends"
        ]
      },
      "patch": {
        "operationId": "updateFriend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFriendDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendDto"
                }
              }
            }
          }
        },
        "summary": "Update a friend",
        "tags": [
          "friends"
        ]
      },
      "delete": {
        "operationId": "deleteFriend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete a friend",
        "tags": [
          "friends"
        ]
      }
    },
    "/friends/{id}/avatar/versions": {
      "get": {
        "operationId": "friendAvatarVersions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarVersionsDto"
                }
              }
            }
          }
        },
        "summary": "A friend's past avatars, newest first",
        "tags": [
          "friends"
        ]
      }
    },
    "/friends/{id}/avatar/versions/{versionId}/restore": {
      "post": {
        "operationId": "restoreFriendAvatarVersion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendDto"
                }
              }
            }
          }
        },
        "summary": "Restore a past avatar as the current one",
        "tags": [
          "friends"
        ]
      }
    },
    "/friends/{id}/connect": {
      "post": {
        "operationId": "connectFriend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendDto"
                }
              }
            }
          }
        },
        "summary": "Add an existing entry as someone you know",
        "tags": [
          "friends"
        ]
      },
      "delete": {
        "operationId": "disconnectFriend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendDto"
                }
              }
            }
          }
        },
        "summary": "Withdraw a pending request, or remove an existing connection",
        "tags": [
          "friends"
        ]
      }
    },
    "/friends/{id}/claim": {
      "post": {
        "operationId": "claimFriend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendDto"
                }
              }
            }
          }
        },
        "summary": "Claim an unassociated friend entry as your own account",
        "tags": [
          "friends"
        ]
      }
    },
...
    "/connections/{userId}": {
      "delete": {
        "operationId": "disconnect",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Remove an existing connection with an account",
        "tags": [
          "connections"
        ]
      }
    },
...
    "/users/search": {
      "get": {
        "operationId": "searchUsers",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Match against name or @handle",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsersDto"
                }
              }
            }
          }
        },
        "summary": "Find accounts to connect with, by name or @handle",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me": {
      "patch": {
        "operationId": "updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Update the current user's account fields (e.g. handle)",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hatohui API",
    "description": "OpenAPI specification for the Hatohui workspace API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "description": "The account's display name — its own chosen name if set, otherwise the name from Google login."
          },
          "handle": {
            "type": "string",
            "nullable": true,
            "description": "Global unique handle (e.g. \"hatohui\"), without the @"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean",
            "description": "Whether this account is the configured admin. Cosmetic only — admin routes enforce their own check and also require the admin key."
          },
          "onboardingStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "COMPLETED",
              "SKIPPED"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "handle",
          "avatarUrl",
          "isAdmin",
          "onboardingStatus"
        ]
      },
      "FriendDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1234567890",
            "description": "Unique friend id"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "birthYear": {
            "type": "number",
            "example": 1998,
            "nullable": true,
            "description": "Birth year. Null when no birthday is recorded, or the viewer is not allowed to see it."
          },
          "birthMonth": {
            "type": "number",
            "example": 5,
            "nullable": true,
            "description": "Birth month (1-12), null when the birthday is not visible"
          },
          "birthDay": {
            "type": "number",
            "example": 14,
            "nullable": true,
            "description": "Birth day (1-31), null when the birthday is not visible"
          },
          "socialMedias": {
            "type": "object",
            "example": {
              "Twitter": "@jane",
              "Instagram": "@jane.doe"
            },
            "nullable": true
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "FRIENDS_ONLY",
              "NONE"
            ],
            "example": "PUBLIC",
            "nullable": true,
            "description": "Who can see the birthday. Null when there is no birthday to govern. The profile itself is always public."
          },
          "avatarUrl": {
            "type": "string",
            "example": "http://localhost:9000/hatohui-dev/avatars/clx1234567890/abc123.jpg",
            "nullable": true,
            "description": "Public URL of the friend's avatar image"
          },
          "addedById": {
            "type": "string",
            "nullable": true,
            "description": "Id of the account that added this entry, if any"
          },
          "handle": {
            "type": "string",
            "example": "janedoe",
            "nullable": true,
            "description": "The profile's global @handle, if it has one"
          },
          "isAssociated": {
            "type": "boolean",
            "description": "Whether this entry is already claimed by an account"
          },
          "isViewerEntry": {
            "type": "boolean",
            "description": "Whether this entry is the requesting viewer's own entry"
          },
          "connectionStatus": {
            "type": "string",
            "enum": [
              "NONE",
              "PENDING_OUTGOING",
              "PENDING_INCOMING",
              "ACCEPTED"
            ],
            "description": "How the viewer stands with the account behind this entry. Always NONE for entries nobody has claimed — there's no account to connect with."
          },
          "canEdit": {
            "type": "boolean",
            "description": "Whether the requesting viewer is allowed to edit/delete this entry"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-07-23T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-07-23T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "birthYear",
          "birthMonth",
          "birthDay",
          "socialMedias",
          "visibility",
          "avatarUrl",
          "addedById",
          "handle",
          "isAssociated",
          "isViewerEntry",
          "connectionStatus",
          "canEdit",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpcomingSectionDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "2026-08",
            "description": "The section's grouping key: YYYY-MM for 'month', the age/year as a string for 'age'/'year', 'unknown' when the underlying field is unset, or 'all' for group 'none'"
          },
          "friends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpcomingFriendDto"
            }
          }
        },
        "required": [
          "key",
          "friends"
        ]
      },
      "PaginatedUpcomingSectionsDto": {
        "type": "object",
        "properties": {
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpcomingSectionDto"
            }
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether another page can be requested"
          }
        },
        "required": [
          "sections",
          "page",
          "pageSize",
          "hasMore"
        ]
      },
      "BirthdaysByMonthDto": {
        "type": "object",
        "properties": {
          "friends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FriendDto"
            }
          }
        },
        "required": [
          "friends"
        ]
      },
      "PaginatedFriendsDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FriendDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
...
      "CreateFriendDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "birthYear": {
            "type": "number",
            "example": 1998
          },
          "birthMonth": {
            "type": "number",
            "example": 5
          },
          "birthDay": {
            "type": "number",
            "example": 14
          },
          "socialMedias": {
            "type": "object",
            "example": {
              "Twitter": "@jane",
              "Instagram": "@jane.doe"
            }
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "FRIENDS_ONLY",
              "NONE"
            ],
            "default": "PUBLIC",
            "description": "Who can see the birthday. Ignored if no date is given."
          },
          "avatarKey": {
            "type": "string",
            "example": "uploads/clx1234567890/abc123.jpg",
            "description": "Object key returned by POST /images/sign, after uploading the avatar file to storage"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateFriendDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "birthYear": {
            "type": "number",
            "example": 1998
          },
          "birthMonth": {
            "type": "number",
            "example": 5
          },
          "birthDay": {
            "type": "number",
            "example": 14
          },
          "socialMedias": {
            "type": "object",
            "example": {
              "Twitter": "@jane",
              "Instagram": "@jane.doe"
            }
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "FRIENDS_ONLY",
              "NONE"
            ],
            "default": "PUBLIC",
            "description": "Who can see the birthday. Ignored if no date is given."
          },
          "avatarKey": {
            "type": "string",
            "example": "uploads/clx1234567890/abc123.jpg",
            "description": "Object key returned by POST /images/sign, after uploading the avatar file to storage"
          }
        }
      },
      "AvatarVersionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx1234567890",
            "description": "Avatar version id"
          },
          "url": {
            "type": "string",
            "example": "http://localhost:9000/hatohui-dev/avatars/clx1234567890/abc123.jpg",
            "description": "Public URL of this past avatar"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-07-23T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "url",
          "createdAt"
        ]
      },
      "AvatarVersionsDto": {
        "type": "object",
        "properties": {
          "versions": {
            "description": "Past avatars, newest first",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AvatarVersionDto"
            }
          }
        },
        "required": [
          "versions"
        ]
      },
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "handle": {
            "type": "string",
            "nullable": true,
            "description": "Global unique handle, without the @"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "handle",
          "avatarUrl"
        ]
      },
...
      "PaginatedUsersDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicUserDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "handle": {
            "type": "string",
            "example": "hatohui",
            "description": "Global unique handle, lowercase letters/digits/underscore, 3-20 chars"
          },
          "displayName": {
            "type": "string",
            "example": "Hatohui",
            "description": "Display name, overriding the name from Google login. Unlike handle, any casing/spacing/accents are fine."
          }
        }
      },
...
    }
  }
}